refactor(cron): split pagination helpers
