+++ src/gateway/server-methods/cron.ts
@@ -22,13 +22,8 @@ import {
 } from "../../cron/delivery-channel-validation.js";
 import { resolveCronDeliveryPreviews } from "../../cron/delivery-preview.js";
 import { assertCronDeliveryInputNonBlankFields } from "../../cron/delivery-target-validation.js";
-import { resolveCronListSnapshotRevision } from "../../cron/list-snapshot-revision.js";
 import { normalizeCronJobCreate, normalizeCronJobPatch } from "../../cron/normalize.js";
 import { applyJobPatch } from "../../cron/service/jobs.js";
-import type {
-  CronListPageOptions,
-  CronListPageResult,
-} from "../../cron/service/list-page-types.js";
 import {
   isInvalidCronSessionTargetIdError,
   resolveCronSessionTargetSessionKey,
@@ -58,9 +53,9 @@ import {
   cronJobMatchesCallerScope,
   cronPatchSessionRefsMatchCaller,
   readCronCallerScope,
-  type CronCallerScope,
 } from "./cron-caller-scope.js";
 import { isCronInvalidRequestError } from "./cron-error-classification.js";
+import { listCronPageForCallerScope } from "./cron-list-caller-scope.js";
 import { cronRunLogPageFilters, filterCronRunLogJobsByAgent } from "./cron-run-log-filters.js";
 import type { GatewayRequestHandlers, RespondFn } from "./types.js";
 
@@ -79,15 +74,6 @@ type CronRunsRequestParams = CronJobIdParams & {
   query?: string;
   sortDir?: "asc" | "desc";
 };
-
-type CronListCallerScopeContext = {
-  cron: {
-    getDefaultAgentId(): string | undefined;
-    listPage(opts?: CronListPageOptions): Promise<CronListPageResult>;
-  };
-};
-
-const CRON_LIST_SCOPED_SNAPSHOT_MAX_ATTEMPTS = 3;
 
 class CronJobConfigRevisionConflictError extends Error {
   constructor(
@@ -147,83 +133,6 @@ function compactCronListJob(job: CronJob) {
     ...(job.state.lastFailureNotificationDeliveryError !== undefined
       ? { lastFailureNotificationDeliveryError: job.state.lastFailureNotificationDeliveryError }
       : {}),
-  };
-}
-
-async function listCronPageForCallerScope({
-  callerScope,
-  context,
-  options,
-}: {
-  callerScope: CronCallerScope;
-  context: CronListCallerScopeContext;
-  options: CronListPageOptions;
-}): Promise<CronListPageResult> {
-  let stableScopedJobs: CronJob[] | undefined;
-  for (let attempt = 0; attempt < CRON_LIST_SCOPED_SNAPSHOT_MAX_ATTEMPTS; attempt += 1) {
-    const scopedJobs: CronJob[] = [];
-    let offset = 0;
-    let snapshotRevision: string | undefined;
-    let snapshotChanged = false;
-
-    for (;;) {
-      const sourcePage = await context.cron.listPage({
-        ...options,
-        // Owner attribution can intentionally differ from a job's execution agent.
-        // Scan source pages, then apply the trusted caller predicate below.
-        agentId: undefined,
-        limit: 200,
-        offset,
-      });
-      if (snapshotRevision && sourcePage.snapshotRevision !== snapshotRevision) {
-        snapshotChanged = true;
-        break;
-      }
-      snapshotRevision = sourcePage.snapshotRevision;
-
-      scopedJobs.push(
-        ...sourcePage.jobs.filter((job) =>
-          cronJobMatchesCallerScope({
-            job,
-            callerScope,
-            defaultAgentId: context.cron.getDefaultAgentId(),
-          }),
-        ),
-      );
-
-      if (
-        !sourcePage.hasMore ||
-        sourcePage.nextOffset === null ||
-        sourcePage.nextOffset <= offset
-      ) {
-        break;
-      }
-      offset = sourcePage.nextOffset;
-    }
-    if (!snapshotChanged && snapshotRevision) {
-      stableScopedJobs = scopedJobs;
-      break;
-    }
-  }
-  if (!stableScopedJobs) {
-    throw new Error("cron.list changed repeatedly while applying caller scope");
-  }
-
-  const total = stableScopedJobs.length;
-  const pageOffset = Math.max(0, Math.min(total, Math.floor(options.offset ?? 0)));
-  const defaultLimit = total === 0 ? 50 : total;
-  const limit = Math.max(1, Math.min(200, Math.floor(options.limit ?? defaultLimit)));
-  const jobs = stableScopedJobs.slice(pageOffset, pageOffset + limit);
-  const nextOffset = pageOffset + jobs.length;
-  return {
-    jobs,
-    // Never expose the source revision: it includes jobs hidden by caller scope.
-    snapshotRevision: resolveCronListSnapshotRevision(stableScopedJobs),
-    total,
-    offset: pageOffset,
-    limit,
-    hasMore: nextOffset < total,
-    nextOffset: nextOffset < total ? nextOffset : null,
   };
 }
